requirements.txt
fastapi
python-dotenv
langchain
langchain-google-genai
langchain-community
langchain-experimental
pydantic
google-generativeai
faiss-cpu
streamlit
